Repeat overly long samples
As a bulk transition is a stop-the-world pause, ds-split won't report the
subsequent sample until after the whole DS has been migrated.  This
means, for instance, a 500 ms pause would only have one sample
associated with it, which looks like a single blip in the viz system.

This patch repeats a sample if it's more than a multiple of the sampling
rate (99ms) to better reflect this.